python/src/requirements.txt
pyaml~=6.0.1
google-cloud-datastore~=2.16.1
